Consolidate CharacterEntry
